forty_two_project/requirements.txt
Django==1.11
django-registration-redux==2.4
django-taggit==0.22.2
pytz==2018.5
